src/physical_commodity_arbitrage_engine/pyproject.toml
[project]
name = "physical-commodity-arbitrage-engine"
version = "0.1.0"
description = "Self-hosted, local Physical Commodity Arbitrage Engine"
authors = [
    { name = "Principal Engineer", email = "engineer@localhost" }
]
requires-python = ">=3.10"
dependencies = [
    "playwright>=1.42.0",
    "streamlit>=1.32.0",
    "python-dotenv>=1.0.1",
    "pandas>=2.2.1",
    "beautifulsoup4>=4.12.3",
    "loguru>=0.7.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/physical_commodity_arbitrage_engine"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
]
